Added shell script for Cygwin LaTeX users
There are various problems with the interaction between scons, pandoc,
and cygwin's version of pdflatex which is a never-ending rabbit hole of
issues of passing env variables, sym-link incompatibilities and path
style incompatibilities.

This script is a hack to work around this until I have the time to fix
it properly.