Replace world map with NOAA atmospheric model
NOAA is a US .gov agency and therefore this image is public domain.

(Wikipedia's exact wording:

This image is a work of the National Oceanic and Atmospheric
Administration, taken or made during the course of an employee's
official duties. As works of the U.S. federal government, all NOAA
images are in the public domain.

)